fix(worker): reject L1 heights that don't fit a block commitment (#227)
`L1DataProvider` handed out Bitcoin heights as `u64`, so the worker
narrowed them with `as u32` before building an `L1BlockCommitment`. A
height above `u32::MAX` wraps silently, and because commitment heights
are part of the persisted sled keys, the worker would then read and
write a different block than the L1 source reported.

Type the trait's heights as `L1Height` instead. That moves the narrowing
into the impls, where the `u64` actually arrives from Bitcoin, and a
checked conversion there fails with `HeightOutOfRange` rather than
truncating.

The genesis comparison now widens the height instead of narrowing
`genesis_height`, since that one is an MMR leaf index and stays `u64`.

Part of SECFIND-224. The same unchecked cast in the RPC server is not
covered here.

+++ crates/worker/src/service.rs
@@ -123,11 +123,12 @@ where
     // height the worker takes from outside; every later height is derived from
     // the parent chain as the STF processes each block.
     let genesis_height = state.genesis_height();
-    let height = state.context.get_l1_block_height(target_blkid)? as u32;
+    let height = state.context.get_l1_block_height(target_blkid)?;
     let target = L1BlockCommitment::new(height, *target_blkid);
 
-    // Ignore blocks before genesis.
-    if height < genesis_height as u32 {
+    // Ignore blocks before genesis. Compared in the wider domain because
+    // `genesis_height` is an MMR leaf index, which is a `u64`.
+    if u64::from(height) < genesis_height {
         warn!(height, "ignoring unexpected L1 block before genesis");
         return Ok(vec![]);
     }
